ref: 소셜 로그인 버튼 로직 수정 및 이름 변경
